Metadata and Ukrainian
Adding metadata to the standard locations will help include
the app in F-Droid. You can translate it to more languages, and
F-Droid will automatically pick it up. I named the ongoing
changelog 2.txt, assuming you will bump the versionCode to 2
with the versionName of 1.1, and tag the release with v1.1